requirements/requirement.txt
# -e git+https://github.com/pcolmant/djangocms-cascade.git@883f39eac05f77d0638df3719bbe5a5bb7b10048#egg=djangocms-cascade
django-cms==3.5.2
# https://github.com/divio/django-cms/archive/develop.zip
djangocms-cascade==0.16.2
# https://github.com/jrief/djangocms-cascade/archive/master.zip
# TODO : remove cmsplugin wich is replaced by djangocms_xxxx
cmsplugin-filer==1.1.3
# django-select2==6.0.2
Django==1.11.15
Pillow==5.3.0
PyYAML==3.13
Unidecode==1.0.22
django-debug-toolbar==1.10.1
diff-match-patch==20121119
django-angular==2.0.3
django-classy-tags==0.8.0
django-crispy-forms==1.7.2
django-easy-select2==1.5.5
django-filer==1.3.2
django-formtools==2.1
django-import-export==1.1.0
django-mptt-admin==0.6.0
django-mptt==0.9.1
django-parler==1.9.2
django-polymorphic==1.3
django-recurrence==1.7.0
django-reversion==3.0.1
django-sekizai==0.10.0
django-treebeard==4.3
djangocms-admin-style==1.2.9
djangocms-attributes-field==0.3.0
djangocms-file==2.0.2
djangocms-link==2.1.2
djangocms-picture==2.0.7
djangocms-text-ckeditor==3.6.0
djangocms-video==2.0.5
djangorestframework==3.9.0
easy-thumbnails==2.5
html5lib==0.9999999
jsonfield==2.0.2
odfpy==1.3.6
olefile==0.45.1
# moved to repanier.packages : openpyxl==1.8.6, with bug fix on worker.worksheet.py lines 103-106
openpyxl==2.5.9
pkg-resources==0.0.0
psycopg2==2.7.5
pyaml==17.8.0
python-dateutil==2.7.3
pytz==2018.5
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.4
tablib==0.11.5
uWSGI==2.0.17.1
unicodecsv==0.14.1
xlrd==1.0.0
xlwt==1.2.0